edit ShowPlan, ShowDataSet
