+++ server/src/main/java/org/apache/iotdb/db/qp/physical/sys/ShowTimeSeriesPlan.java
@@ -28,40 +28,29 @@ public class ShowTimeSeriesPlan extends ShowPlan {
 
   // path can be root, root.*  root.*.*.a etc.. if the wildcard is not at the tail, then each
   // * wildcard can only match one level, otherwise it can match to the tail.
-  private PartialPath path;
   private boolean isContains;
   private String key;
   private String value;
-  private int limit = 0;
-  private int offset = 0;
+
   // if is true, the result will be sorted according to the inserting frequency of the timeseries
   private boolean orderByHeat;
 
-  private boolean hasLimit;
 
   public ShowTimeSeriesPlan(PartialPath path) {
-    super(ShowContentType.TIMESERIES);
-    this.path = path;
+    super(ShowContentType.TIMESERIES, path);
   }
 
   public ShowTimeSeriesPlan(PartialPath path, boolean isContains, String key, String value, int limit,
-      int offset, boolean orderByHeat) {
-    super(ShowContentType.TIMESERIES);
-    this.path = path;
+      int offset, boolean orderByHeat, int fetchSize) {
+    super(ShowContentType.TIMESERIES, path, limit, offset, fetchSize);
     this.isContains = isContains;
     this.key = key;
     this.value = value;
-    this.limit = limit;
-    this.offset = offset;
     this.orderByHeat = orderByHeat;
   }
 
   public ShowTimeSeriesPlan() {
     super(ShowContentType.TIMESERIES);
-  }
-
-  public PartialPath getPath() {
-    return this.path;
   }
 
   public boolean isContains() {
@@ -76,22 +65,6 @@ public class ShowTimeSeriesPlan extends ShowPlan {
     return value;
   }
 
-  public int getLimit() {
-    return limit;
-  }
-
-  public void setLimit(int limit) {
-    this.limit = limit;
-  }
-
-  public int getOffset() {
-    return offset;
-  }
-
-  public void setOffset(int offset) {
-    this.offset = offset;
-  }
-
   public boolean isOrderByHeat() {
     return orderByHeat;
   }
@@ -100,13 +73,7 @@ public class ShowTimeSeriesPlan extends ShowPlan {
     this.orderByHeat = orderByHeat;
   }
 
-  public boolean hasLimit() {
-    return hasLimit;
-  }
 
-  public void setHasLimit(boolean hasLimit) {
-    this.hasLimit = hasLimit;
-  }
 
   @Override
   public void serialize(DataOutputStream outputStream) throws IOException {
