+++ server/src/main/java/org/apache/iotdb/db/service/TSServiceImpl.java
@@ -542,26 +542,6 @@ public class TSServiceImpl implements TSIService.Iface, ServerContext {
         fetchSize = DEFAULT_FETCH_SIZE;
       }
 
-      if (plan instanceof ShowTimeSeriesPlan) {
-        //If the user does not pass the limit, then set limit = fetchSize and haslimit=false,else set haslimit = true
-        if (((ShowTimeSeriesPlan) plan).getLimit() == 0) {
-          ((ShowTimeSeriesPlan) plan).setLimit(fetchSize);
-          ((ShowTimeSeriesPlan) plan).setHasLimit(false);
-        } else {
-          ((ShowTimeSeriesPlan) plan).setHasLimit(true);
-        }
-      }
-
-      if (plan instanceof ShowDevicesPlan) {
-        //If the user does not pass the limit, then set limit = fetchSize and haslimit=false,else set haslimit = true
-        if (((ShowDevicesPlan) plan).getLimit() == 0) {
-          ((ShowDevicesPlan) plan).setLimit(fetchSize);
-          ((ShowDevicesPlan) plan).setHasLimit(false);
-        } else {
-          ((ShowDevicesPlan) plan).setHasLimit(true);
-        }
-      }
-
       if (plan instanceof QueryPlan && !((QueryPlan) plan).isAlignByTime()) {
         if (plan.getOperatorType() == OperatorType.AGGREGATION) {
           throw new QueryProcessException("Aggregation doesn't support disable align clause.");
